Sync global data with widget and global editor fields - Introduce Open Api module (for all communcation with the api) - Add sync logic to openstad-widgets - Add sync logic to apostrophe-global widget - Implement sync fields in idea-form (minTitle, maxTitle etc) - Use getIdeas from openstad-api in apostrophe-page
