+++ lib/modules/apostrophe-pages/lib/load-ideas.js
@@ -41,7 +41,6 @@ module.exports =  function (req, res, next) {
       ideas = cache.get(cacheKey);
     }
 
-
     // if cacheIdeas is turned on, get ideas from cache
     if (ideas && ideas.length > 0) {
       req.data.ideas = ideas;
