+++ lib/modules/idea-form-widgets/index.js
@@ -25,39 +25,14 @@ module.exports = {
   extend: 'map-widgets',
   label: 'Idea form',
   addFields: fields,
+  openstadApiConfigSync: true,
   afterConstruct: function(self) {
-      
-      const superLoad = self.load;
-      self.load = function(req, widgets, next) {
-          widgets.forEach((widget) => {
 
-              // Todo: get data from api
-              //widget.minTitle = req.data.global.siteConfig.ideas.titleMinLength;
-              //widget.maxTitle = req.data.global.siteConfig.ideas.titleMaxLength;
-
-              return superLoad(req, widgets, next);
-          });
-      };
 
   },
   construct: function(self, options) {
 
-      self.on('apostrophe-docs:afterSave', 'syncApi', async (req, doc, options) => {
-          console.log('test after save')
-          console.log(doc);
-          if (doc.type !== 'idea-form') {
-              return;
-          }
-
-          // Todo: sync data with api
-          console.log(doc.type);
-
-          console.log(doc.minTitle);
-          console.log(doc.maxTitle);
-
-      });
-
-   options.arrangeFields = (options.arrangeFields || []).concat([
+      options.arrangeFields = (options.arrangeFields || []).concat([
      {
        name: 'general',
        label: 'Algemeen',
@@ -128,21 +103,13 @@ module.exports = {
     /** add config **/
     const superLoad = self.load;
 
-    self.load = function(req, widgets, next) {
+    self.load = async function(req, widgets, next) {
         const styles = self.apos.settings.getOption(req, 'openStadMap').defaults.styles;
         const globalData = req.data.global;
 
-	      widgets.forEach((widget) => {
+        widgets.forEach(async (widget) => {
 	          const siteConfig = req.data.global.siteConfig;
     				widget.siteConfig = {
-    					ideas: {
-    						titleMinLength: ( siteConfig && siteConfig.ideas && siteConfig.ideas.titleMinLength ) || 10,
-    						titleMaxLength: ( siteConfig && siteConfig.ideas && siteConfig.ideas.titleMaxLength ) || 50,
-    						summaryMinLength: ( siteConfig && siteConfig.ideas && siteConfig.ideas.summaryMinLength ) || 20,
-    						summaryMaxLength: ( siteConfig && siteConfig.ideas && siteConfig.ideas.summaryMaxLength ) || 140,
-    						descriptionMinLength: ( siteConfig && siteConfig.ideas && siteConfig.ideas.descriptionMinLength ) || 140,
-    						descriptionMaxLength: ( siteConfig && siteConfig.ideas && siteConfig.ideas.descriptionMaxLength ) || 5000,
-    					},
     					openstadMap: {
     						polygon: ( siteConfig && siteConfig.openstadMap && siteConfig.openstadMap.polygon ) || undefined,
     					},
@@ -172,7 +139,7 @@ module.exports = {
   		});
 
 			superLoad(req, widgets, next);
-		}
+		};
 
 
    /**
