+++ lib/modules/idea-form-widgets/lib/fields.js
@@ -210,11 +210,15 @@ const fields = [
         name: 'minTitle',
         type: 'integer',
         label: 'Minimum characters for Title',
+        apiSyncField: 'ideas.titleMinLength',
+        help: 'This field will also update the global settings field (ideas.titleMinLength) in the api'
     },
     {
         name: 'maxTitle',
         type: 'integer',
         label: 'Maximum characters for Title',
+        apiSyncField: 'ideas.titleMaxLength',
+        help: 'This field will also update the config settings field (ideas.titleMinLength) in the api'
     },
     {
        name: 'labelSummary',
@@ -257,6 +261,21 @@ const fields = [
         ]
     },
     {
+        name: 'minSummary',
+        type: 'integer',
+        label: 'Minimum characters for Summary',
+        apiSyncField: 'ideas.summaryMinLength',
+        help: 'This field will also update the global settings field (ideas.summaryMinLength) in the api'
+    },
+    {
+        name: 'maxSummary',
+        type: 'integer',
+        label: 'Minimum characters for Summary',
+        apiSyncField: 'ideas.summaryMaxLength',
+        help: 'This field will also update the global settings field (ideas.summaryMaxLength) in the api'
+    },
+
+    {
        name: 'labelDescription',
        type: 'string',
        label: 'Label for Description',
@@ -296,6 +315,20 @@ const fields = [
                 label: "No"
             },
         ]
+    },
+    {
+        name: 'minDescription',
+        type: 'integer',
+        label: 'Minimum characters for Description',
+        apiSyncField: 'ideas.descriptionMinLength',
+        help: 'This field will also update the global settings field (ideas.descriptionMinLength) in the api'
+    },
+    {
+        name: 'maxDescription',
+        type: 'integer',
+        label: 'Minimum characters for Description',
+        apiSyncField: 'ideas.descriptionMaxLength',
+        help: 'This field will also update the global settings field (ideas.descriptionMaxLength) in the api'
     },
     {
        name: 'labelImages',
