+++ lib/modules/openstad-api/lib/api.js
@@ -2,48 +2,65 @@ const request = require('request-promise');
 
 module.exports = (self, options) => {
 
-    self.ideas = null;
+  self.init = (req) => {
+    const siteConfig = self.apos.settings.getOption(req, 'siteConfig');
+    self.siteId = siteConfig.id;
+    self.apiUrl = process.env.INTERNAL_API_URL || self.apos.settings.getOption(req, 'apiUrl');
+    self.sessionJwt = req.session.jwt;
+  };
 
-    self.getSite = (siteId) => {
+  self.setApiConfigValue = (config, field, value) => {
+    self.setObjectValue(config, field, value);
+  };
 
-    };
+  self.getApiConfigValue = (config, field) => {
+    return self.setObjectValue(config, field);
+  };
 
-    self.getAllIdeas = async (req, siteId, sort) => {
-        const apiUrl = process.env.INTERNAL_API_URL || req.data.apiUrl;
-        if(self.ideas) {
-            console.log('already have the ideas promise');
-            return self.ideas;
-        }
+  self.updateSiteConfig = async (siteConfig, item, apiSyncFields) => {
 
-        const options = {
-            uri: `${apiUrl}/api/site/${siteId}/idea?sort=${sort}&includeVoteCount=1&includeUserVote=1`,
-            headers: {
-                'Accept': 'application/json',
-            },
-            json: true
-        };
+    apiSyncFields.forEach(field => {
+      const value = self.getFieldValue(item, field);
+      self.setApiConfigValue(siteConfig, field.apiSyncField, value);
+    });
 
-        self.ideas = request(options);
+    return self.updateSite({config: siteConfig});
+  };
 
-        return self.ideas;
-    };
+  self.getOptions = (additionalHeaders) => {
+    return Object.assign({
+      headers: {
+        'Accept': 'application/json',
+      },
+      json: true
+    }, additionalHeaders);
+  };
 
-    self.updateSite = async (req, siteId, data) => {
+  self.getOptionsWithAuth = (additionalHeaders) => {
+    return self.getOptions(Object.assign({
+      headers: {
+        'Accept': 'application/json',
+        'X-Authorization': `Bearer ${self.sessionJwt}`
+      }
+    }, additionalHeaders));
+  };
 
-        const apiUrl = self.apos.settings.getOption(req, 'apiUrl');
+  self.getAllIdeas = async (req, siteId, sort) => {
+    const options = self.getOptions({
+      uri: `${self.apiUrl}/api/site/${siteId}/idea?sort=${sort}&includeVoteCount=1&includeUserVote=1`,
+    });
 
-        const options = {
-            method: 'PUT',
-            uri:  apiUrl + `/api/site/${siteId}`,
-            headers: {
-                'Accept': 'application/json',
-                "X-Authorization" : ` Bearer ${req.session.jwt}`,
-            },
-            body: data,
-            json: true // Automatically parses the JSON string in the response
-        };
+    return request(options);
+  };
 
-        console.log(options);
-        return request(options);
-    };
+  self.updateSite = async (data) => {
+    const options = self.getOptionsWithAuth({
+      method: 'PUT',
+      uri: `${self.apiUrl}/api/site/${self.siteId}`,
+      body: data,
+    });
+
+
+    return request(options);
+  };
 };
